+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 Changelog
 =========
 
+- #2471 Enable **matplotlib** color abbreviations to be passed as `color`, `style_kwds` and `missing_kwds` paramters to `explore()`.  For example `r` for `red`.   [matplotlib colors](https://matplotlib.org/2.0.1/api/colors_api.html#matplotlib.colors.is_color_like)
+
 Version 0.11 (June 20, 2022)
 ----------------------------